added better error handling

+++ commands/secret/amongus.js
@@ -13,24 +13,29 @@ module.exports = {
         amongusers.toString().split(" ");
         if(amongusers.includes(message.author.id.toString())) {
             if (message.member.voice.channel) {
-                const connection = await message.member.voice.channel.join();
+                const connection = await message.member.voice.channel.join()
+                    .catch(err => {
+                        message.channel.send("An error occured joining the channel.");
+                        console.log(err.message)
+                    });
+                if(connection) {
+                    // Create a dispatcher
+                    const dispatcher = connection.play('audio/amongus.mp3',  { volume: 0.25 });
 
-                // Create a dispatcher
-                const dispatcher = connection.play('audio/amongus.mp3',  { volume: 0.25 });
+                    dispatcher.on('start', () => {
+                        console.log('amongus.mp3 is now playing!');
+                    });
 
-                dispatcher.on('start', () => {
-                    console.log('amongus.mp3 is now playing!');
-                });
+                    dispatcher.on('finish', () => {
+                        console.log('amongus.mp3 has finished playing!');
+                        dispatcher.destroy();
+                        connection.disconnect();
+                        message.delete().catch(err => console.log(err.message));
+                    });
 
-                dispatcher.on('finish', () => {
-                    console.log('amongus.mp3 has finished playing!');
-                    dispatcher.destroy();
-                    connection.disconnect();
-                    message.delete().catch(err => console.log(err.message));
-                });
-
-                // Always remember to handle errors appropriately!
-                dispatcher.on('error', console.error);
+                    // Always remember to handle errors appropriately!
+                    dispatcher.on('error', console.error);
+                }
             } else {
                 message.channel.send("you need to be in a vc dumbass!").catch(err => console.log(err.message));
             }
